Matchers functions moverd to network_matchers.rs.
