+++ src/filters/mod.rs
@@ -1,5 +1,7 @@
 //! Contains representations and standalone behaviors of individual filter rules.
 
 mod abstract_network;
+mod network_matchers;
+
 pub mod cosmetic;
 pub mod network;
