+++ src/filters/network.rs
@@ -1,7 +1,7 @@
 //! Filters that take effect at the network request level, including blocking and response
 //! modification.
 
-use memchr::{memchr as find_char, memmem};
+use memchr::memchr as find_char;
 use once_cell::sync::Lazy;
 use regex::Regex;
 use serde::{Deserialize, Serialize};
@@ -982,7 +982,22 @@ pub trait NetworkMatchable {
 
 impl NetworkMatchable for NetworkFilter {
     fn matches(&self, request: &request::Request, regex_manager: &mut RegexManager) -> bool {
-        check_options(self, request) && check_pattern(self, request, regex_manager)
+        use crate::filters::network_matchers::{check_options, check_pattern};
+        check_options(
+            self.mask,
+            self.opt_domains.as_deref(),
+            self.opt_domains_union,
+            self.opt_not_domains.as_deref(),
+            self.opt_not_domains_union,
+            request,
+        ) && check_pattern(
+            self.mask,
+            self.filter.iter(),
+            self.hostname.as_deref(),
+            (self as *const NetworkFilter) as u64,
+            request,
+            regex_manager,
+        )
     }
 
     #[cfg(test)]
@@ -1051,439 +1066,6 @@ fn check_is_regex(filter: &str) -> bool {
     let separator_index = find_char(b'^', filter.as_bytes());
     start_index.is_some() || separator_index.is_some()
 }
-
-/// Handle hostname anchored filters, given 'hostname' from ||hostname and
-/// request's hostname, check if there is a match. This is tricky because
-/// filters authors rely and different assumption. We can have prefix of suffix
-/// matches of anchor.
-fn is_anchored_by_hostname(
-    filter_hostname: &str,
-    hostname: &str,
-    wildcard_filter_hostname: bool,
-) -> bool {
-    let filter_hostname_len = filter_hostname.len();
-    // Corner-case, if `filterHostname` is empty, then it's a match
-    if filter_hostname_len == 0 {
-        return true;
-    }
-    let hostname_len = hostname.len();
-
-    if filter_hostname_len > hostname_len {
-        // `filterHostname` cannot be longer than actual hostname
-        false
-    } else if filter_hostname_len == hostname_len {
-        // If they have the same len(), they should be equal
-        filter_hostname == hostname
-    } else if let Some(match_index) = memmem::find(hostname.as_bytes(), filter_hostname.as_bytes())
-    {
-        if match_index == 0 {
-            // `filter_hostname` is a prefix of `hostname` and needs to match full a label.
-            //
-            // Examples (filter_hostname, hostname):
-            //   * (foo, foo.com)
-            //   * (sub.foo, sub.foo.com)
-            wildcard_filter_hostname
-                || filter_hostname.ends_with('.')
-                || hostname[filter_hostname_len..].starts_with('.')
-        } else if match_index == hostname_len - filter_hostname_len {
-            // `filter_hostname` is a suffix of `hostname`.
-            //
-            // Examples (filter_hostname, hostname):
-            //    * (foo.com, sub.foo.com)
-            //    * (com, foo.com)
-            filter_hostname.starts_with('.') || hostname[match_index - 1..].starts_with('.')
-        } else {
-            // `filter_hostname` is infix of `hostname` and needs match full labels
-            (wildcard_filter_hostname
-                || filter_hostname.ends_with('.')
-                || hostname[filter_hostname_len..].starts_with('.'))
-                && (filter_hostname.starts_with('.')
-                    || hostname[match_index - 1..].starts_with('.'))
-        }
-    } else {
-        // No match
-        false
-    }
-}
-
-fn get_url_after_hostname<'a>(url: &'a str, hostname: &str) -> &'a str {
-    let start =
-        memmem::find(url.as_bytes(), hostname.as_bytes()).unwrap_or(url.len() - hostname.len());
-    &url[start + hostname.len()..]
-}
-
-// ---------------------------------------------------------------------------
-// Filter matching
-// ---------------------------------------------------------------------------
-
-// pattern
-fn check_pattern_plain_filter_filter(filter: &NetworkFilter, request: &request::Request) -> bool {
-    let request_url = request.get_url(filter.match_case());
-    match &filter.filter {
-        FilterPart::Empty => true,
-        FilterPart::Simple(f) => memmem::find(request_url.as_bytes(), f.as_bytes()).is_some(),
-        FilterPart::AnyOf(filters) => {
-            for f in filters {
-                if memmem::find(request_url.as_bytes(), f.as_bytes()).is_some() {
-                    return true;
-                }
-            }
-            false
-        }
-    }
-}
-
-// pattern|
-fn check_pattern_right_anchor_filter(filter: &NetworkFilter, request: &request::Request) -> bool {
-    let request_url = request.get_url(filter.match_case());
-    match &filter.filter {
-        FilterPart::Empty => true,
-        FilterPart::Simple(f) => request_url.ends_with(f),
-        FilterPart::AnyOf(filters) => {
-            for f in filters {
-                if request_url.ends_with(f) {
-                    return true;
-                }
-            }
-            false
-        }
-    }
-}
-
-// |pattern
-fn check_pattern_left_anchor_filter(filter: &NetworkFilter, request: &request::Request) -> bool {
-    let request_url = request.get_url(filter.match_case());
-    match &filter.filter {
-        FilterPart::Empty => true,
-        FilterPart::Simple(f) => request_url.starts_with(f),
-        FilterPart::AnyOf(filters) => {
-            for f in filters {
-                if request_url.starts_with(f) {
-                    return true;
-                }
-            }
-            false
-        }
-    }
-}
-
-// |pattern|
-fn check_pattern_left_right_anchor_filter(
-    filter: &NetworkFilter,
-    request: &request::Request,
-) -> bool {
-    let request_url = request.get_url(filter.match_case());
-    match &filter.filter {
-        FilterPart::Empty => true,
-        FilterPart::Simple(f) => &request_url == f,
-        FilterPart::AnyOf(filters) => {
-            for f in filters {
-                if &request_url == f {
-                    return true;
-                }
-            }
-            false
-        }
-    }
-}
-
-// pattern*^
-fn check_pattern_regex_filter_at(
-    filter: &NetworkFilter,
-    request: &request::Request,
-    start_from: usize,
-    regex_manager: &mut RegexManager,
-) -> bool {
-    let request_url = request.get_url(filter.match_case());
-    regex_manager.matches(filter, &request_url[start_from..])
-}
-
-fn check_pattern_regex_filter(
-    filter: &NetworkFilter,
-    request: &request::Request,
-    regex_manager: &mut RegexManager,
-) -> bool {
-    check_pattern_regex_filter_at(filter, request, 0, regex_manager)
-}
-
-// ||pattern*^
-fn check_pattern_hostname_anchor_regex_filter(
-    filter: &NetworkFilter,
-    request: &request::Request,
-    regex_manager: &mut RegexManager,
-) -> bool {
-    let request_url = request.get_url(filter.match_case());
-    filter
-        .hostname
-        .as_ref()
-        .map(|hostname| {
-            if is_anchored_by_hostname(
-                hostname,
-                &request.hostname,
-                filter.mask.contains(NetworkFilterMask::IS_HOSTNAME_REGEX),
-            ) {
-                check_pattern_regex_filter_at(
-                    filter,
-                    request,
-                    memmem::find(request_url.as_bytes(), hostname.as_bytes()).unwrap_or_default()
-                        + hostname.len(),
-                    regex_manager,
-                )
-            } else {
-                false
-            }
-        })
-        .unwrap_or_else(|| unreachable!()) // no match if filter has no hostname - should be unreachable
-}
-
-// ||pattern|
-fn check_pattern_hostname_right_anchor_filter(
-    filter: &NetworkFilter,
-    request: &request::Request,
-) -> bool {
-    filter
-        .hostname
-        .as_ref()
-        .map(|hostname| {
-            if is_anchored_by_hostname(
-                hostname,
-                &request.hostname,
-                filter.mask.contains(NetworkFilterMask::IS_HOSTNAME_REGEX),
-            ) {
-                match &filter.filter {
-                    // In this specific case it means that the specified hostname should match
-                    // at the end of the hostname of the request. This allows to prevent false
-                    // positive like ||foo.bar which would match https://foo.bar.baz where
-                    // ||foo.bar^ would not.
-                    FilterPart::Empty => {
-                        request.hostname.len() == hostname.len()        // if lengths are equal, hostname equality is implied by anchoring check
-                            || request.hostname.ends_with(hostname)
-                    }
-                    _ => check_pattern_right_anchor_filter(filter, request),
-                }
-            } else {
-                false
-            }
-        })
-        .unwrap_or_else(|| unreachable!()) // no match if filter has no hostname - should be unreachable
-}
-
-// |||pattern|
-fn check_pattern_hostname_left_right_anchor_filter(
-    filter: &NetworkFilter,
-    request: &request::Request,
-) -> bool {
-    // Since this is not a regex, the filter pattern must follow the hostname
-    // with nothing in between. So we extract the part of the URL following
-    // after hostname and will perform the matching on it.
-
-    filter
-        .hostname
-        .as_ref()
-        .map(|hostname| {
-            if is_anchored_by_hostname(
-                hostname,
-                &request.hostname,
-                filter.mask.contains(NetworkFilterMask::IS_HOSTNAME_REGEX),
-            ) {
-                let request_url = request.get_url(filter.match_case());
-                match &filter.filter {
-                    // if no filter, we have a match
-                    FilterPart::Empty => true,
-                    // Since it must follow immediatly after the hostname and be a suffix of
-                    // the URL, we conclude that filter must be equal to the part of the
-                    // url following the hostname.
-                    FilterPart::Simple(f) => get_url_after_hostname(&request_url, hostname) == f,
-                    FilterPart::AnyOf(filters) => {
-                        let url_after_hostname = get_url_after_hostname(&request_url, hostname);
-                        for f in filters {
-                            if url_after_hostname == f {
-                                return true;
-                            }
-                        }
-                        false
-                    }
-                }
-            } else {
-                false
-            }
-        })
-        .unwrap_or_else(|| unreachable!()) // no match if filter has no hostname - should be unreachable
-}
-
-// ||pattern + left-anchor => This means that a plain pattern needs to appear
-// exactly after the hostname, with nothing in between.
-fn check_pattern_hostname_left_anchor_filter(
-    filter: &NetworkFilter,
-    request: &request::Request,
-) -> bool {
-    filter
-        .hostname
-        .as_ref()
-        .map(|hostname| {
-            if is_anchored_by_hostname(
-                hostname,
-                &request.hostname,
-                filter.mask.contains(NetworkFilterMask::IS_HOSTNAME_REGEX),
-            ) {
-                let request_url = request.get_url(filter.match_case());
-                match &filter.filter {
-                    // if no filter, we have a match
-                    FilterPart::Empty => true,
-                    // Since this is not a regex, the filter pattern must follow the hostname
-                    // with nothing in between. So we extract the part of the URL following
-                    // after hostname and will perform the matching on it.
-                    FilterPart::Simple(f) => {
-                        get_url_after_hostname(&request_url, hostname).starts_with(f)
-                    }
-                    FilterPart::AnyOf(filters) => {
-                        let url_after_hostname = get_url_after_hostname(&request_url, hostname);
-                        for f in filters {
-                            if url_after_hostname.starts_with(f) {
-                                return true;
-                            }
-                        }
-                        false
-                    }
-                }
-            } else {
-                false
-            }
-        })
-        .unwrap_or_else(|| unreachable!()) // no match if filter has no hostname - should be unreachable
-}
-
-// ||pattern
-fn check_pattern_hostname_anchor_filter(
-    filter: &NetworkFilter,
-    request: &request::Request,
-) -> bool {
-    filter
-        .hostname
-        .as_ref()
-        .map(|hostname| {
-            if is_anchored_by_hostname(
-                hostname,
-                &request.hostname,
-                filter.mask.contains(NetworkFilterMask::IS_HOSTNAME_REGEX),
-            ) {
-                let request_url = request.get_url(filter.match_case());
-                match &filter.filter {
-                    // if no filter, we have a match
-                    FilterPart::Empty => true,
-                    // Filter hostname does not necessarily have to be a full, proper hostname, part of it can be lumped together with the URL
-                    FilterPart::Simple(f) => {
-                        get_url_after_hostname(&request_url, hostname).contains(f)
-                    }
-                    FilterPart::AnyOf(filters) => {
-                        let url_after_hostname = get_url_after_hostname(&request_url, hostname);
-                        for f in filters {
-                            if url_after_hostname.contains(f) {
-                                return true;
-                            }
-                        }
-                        false
-                    }
-                }
-            } else {
-                false
-            }
-        })
-        .unwrap_or_else(|| unreachable!()) // no match if filter has no hostname - should be unreachable
-}
-
-/// Efficiently checks if a certain network filter matches against a network
-/// request.
-fn check_pattern(
-    filter: &NetworkFilter,
-    request: &request::Request,
-    regex_manager: &mut RegexManager,
-) -> bool {
-    if filter.is_hostname_anchor() {
-        if filter.is_regex() {
-            check_pattern_hostname_anchor_regex_filter(filter, request, regex_manager)
-        } else if filter.is_right_anchor() && filter.is_left_anchor() {
-            check_pattern_hostname_left_right_anchor_filter(filter, request)
-        } else if filter.is_right_anchor() {
-            check_pattern_hostname_right_anchor_filter(filter, request)
-        } else if filter.is_left_anchor() {
-            check_pattern_hostname_left_anchor_filter(filter, request)
-        } else {
-            check_pattern_hostname_anchor_filter(filter, request)
-        }
-    } else if filter.is_regex() || filter.is_complete_regex() {
-        check_pattern_regex_filter(filter, request, regex_manager)
-    } else if filter.is_left_anchor() && filter.is_right_anchor() {
-        check_pattern_left_right_anchor_filter(filter, request)
-    } else if filter.is_left_anchor() {
-        check_pattern_left_anchor_filter(filter, request)
-    } else if filter.is_right_anchor() {
-        check_pattern_right_anchor_filter(filter, request)
-    } else {
-        check_pattern_plain_filter_filter(filter, request)
-    }
-}
-
-fn check_options(filter: &NetworkFilter, request: &request::Request) -> bool {
-    // Bad filter never matches
-    if filter.is_badfilter() {
-        return false;
-    }
-    // We first discard requests based on type, protocol and party. This is really
-    // cheap and should be done first.
-    if !filter.check_cpt_allowed(&request.request_type)
-        || (request.is_https && !filter.for_https())
-        || (request.is_http && !filter.for_http())
-        || (!filter.first_party() && !request.is_third_party)
-        || (!filter.third_party() && request.is_third_party)
-    {
-        return false;
-    }
-
-    // Source URL must be among these domains to match
-    if let Some(included_domains) = filter.opt_domains.as_ref() {
-        if let Some(source_hashes) = request.source_hostname_hashes.as_ref() {
-            // If the union of included domains is recorded
-            if let Some(included_domains_union) = filter.opt_domains_union {
-                // If there isn't any source hash that matches the union, there's no match at all
-                if source_hashes
-                    .iter()
-                    .all(|h| h & included_domains_union != *h)
-                {
-                    return false;
-                }
-            }
-            if source_hashes
-                .iter()
-                .all(|h| !utils::bin_lookup(included_domains, *h))
-            {
-                return false;
-            }
-        }
-    }
-
-    if let Some(excluded_domains) = filter.opt_not_domains.as_ref() {
-        if let Some(source_hashes) = request.source_hostname_hashes.as_ref() {
-            // If the union of excluded domains is recorded
-            if let Some(excluded_domains_union) = filter.opt_not_domains_union {
-                // If there's any source hash that matches the union, check the actual values
-                if source_hashes.iter().any(|h| {
-                    (h & excluded_domains_union == *h) && utils::bin_lookup(excluded_domains, *h)
-                }) {
-                    return false;
-                }
-            } else if source_hashes
-                .iter()
-                .any(|h| utils::bin_lookup(excluded_domains, *h))
-            {
-                return false;
-            }
-        }
-    }
-
-    true
-}
-
 #[cfg(test)]
 #[path = "../../tests/unit/filters/network.rs"]
 mod unit_tests;
